# src/main/java/com/datn/warehousemgmt/utils/PageDTO.java
package com.datn.warehousemgmt.utils;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class PageDTO {
    private Integer page;
    private Integer size;
    String sortBy;
    String sortType;
}