# swagger.json
{"openapi":"3.0.1","info":{"title":"Warehouse Management API Description","description":"API documents","license":{"name":"Apache 2.0","url":"https://springdoc.org"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8134"}],"security":[{"bearerAuth":[]}],"paths":{"/api/users/{uId}":{"get":{"tags":["API quản lý người dùng"],"summary":"Thông tin user","operationId":"getUserById","parameters":[{"name":"uId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"put":{"tags":["API quản lý người dùng"],"summary":"Cập nhật thông tin user","operationId":"updateUser","parameters":[{"name":"uId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceResponse"}}}}}},"delete":{"tags":["API quản lý người dùng"],"summary":"Xóa thông tin user","operationId":"deleteUser","parameters":[{"name":"uId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceResponse"}}}}}}},"/api/product":{"put":{"tags":["API sản phẩm"],"summary":"Cập nhật thông tin sản phẩm","operationId":"updateProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"post":{"tags":["API sản phẩm"],"summary":"Tạo sản phẩm mới","operationId":"createProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["API sản phẩm"],"summary":"Xóa sản phẩm","operationId":"deleteProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/productLog/change-status":{"put":{"tags":["API log nhập/xuất kho"],"summary":"Thay đổi trạng thái log","operationId":"changeLogStatus","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductLogDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/batch":{"put":{"tags":["API Lô hàng"],"summary":"Cập nhật thông tin lô hàng","operationId":"update","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceResponse"}}}}}}},"/api/warehouse/import-goods":{"post":{"tags":["API quản lý kho"],"summary":"Nhập kho","operationId":"importGoods","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RfidDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/warehouse/export-goods":{"post":{"tags":["API quản lý kho"],"summary":"Xuất kho","operationId":"exportGoods","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RfidDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/users":{"post":{"tags":["API quản lý người dùng"],"summary":"Tạo user mới","operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceResponse"}}}}}}},"/api/users/get-more":{"post":{"tags":["API quản lý người dùng"],"summary":"Lấy danh sách/Tìm kiếm thông tin user","operationId":"getAllUsers","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSearchRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceResponse"}}}}}}},"/api/upload/sku-image":{"post":{"tags":["API upload file"],"summary":"Upload ảnh cho mặt hàng","operationId":"uploadAvt","requestBody":{"content":{"application/json":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/upload/excel-file":{"post":{"tags":["API upload file"],"summary":"Upload excel file","operationId":"uploadExcel","parameters":[{"name":"file","in":"query","required":true,"schema":{"type":"string","format":"binary"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/productLog/find-log":{"post":{"tags":["API log nhập/xuất kho"],"summary":"Tìm kiếm, lấy danh sách log","operationId":"findLog","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductLogSearchRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/product/import":{"post":{"tags":["API sản phẩm"],"summary":"Import sản phẩm từ file csv","operationId":"importProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImportRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/product/get":{"post":{"tags":["API sản phẩm"],"summary":"Lấy danh sách sản phẩm","operationId":"getProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/esp/rfid":{"post":{"tags":["API giao tiếp với phần cứng"],"summary":"Đọc thông tin thẻ RFID","operationId":"processRFID","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/batch/findBatch":{"post":{"tags":["API Lô hàng"],"summary":"Tìm kiếm thông tin lô hàng","operationId":"findBatch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/authentication/login":{"post":{"tags":["API Authentication"],"summary":"Đăng nhập","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/productLog/{id}":{"get":{"tags":["API log nhập/xuất kho"],"summary":"Chi tiết log","operationId":"getLog","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/productLog/export":{"get":{"tags":["API log nhập/xuất kho"],"summary":"Xuất báo cáo","operationId":"exportLog","parameters":[{"name":"request","in":"query","required":true,"schema":{"$ref":"#/components/schemas/ExportLogRequest"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/product/{id}":{"get":{"tags":["API sản phẩm"],"summary":"Lấy thông tin sản phẩm theo SKU code","operationId":"getProduct_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/product/genId":{"get":{"tags":["API sản phẩm"],"summary":"API sinh mã sản phẩm","operationId":"generateId","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/batch/{id}":{"get":{"tags":["API Lô hàng"],"summary":"Lấy thông tin lô hàng theo mã","operationId":"getBatchById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceResponse"}}}}}}}},"components":{"schemas":{"UserDTO":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"birthday":{"type":"string","format":"date-time"},"username":{"type":"string"},"password":{"type":"string"},"phoneNumber":{"type":"string"},"gender":{"type":"integer","format":"int32"}}},"ServiceResponse":{"type":"object","properties":{"message":{"type":"string"},"code":{"type":"integer","format":"int32"},"data":{"type":"object"},"totalPage":{"type":"integer","format":"int32"},"totalItem":{"type":"integer","format":"int64"},"currentPage":{"type":"integer","format":"int32"}}},"PageDTO":{"type":"object","properties":{"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"sortBy":{"type":"string"},"sortType":{"type":"string"}}},"ProductRequest":{"type":"object","properties":{"skuCode":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"imageLink":{"type":"string"},"categoryIds":{"type":"array","items":{"type":"integer","format":"int64"}},"pageDTO":{"$ref":"#/components/schemas/PageDTO"},"search":{"type":"string"}}},"Packet":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"rfidTag":{"type":"string"},"quantity":{"type":"integer","format":"int32"},"batchId":{"type":"integer","format":"int64"},"status":{"type":"string"}}},"ProductLogDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"batchProductId":{"type":"integer","format":"int64"},"supplierName":{"type":"string"},"quantity":{"type":"integer","format":"int32"},"action":{"type":"string"},"status":{"type":"string"},"merchantId":{"type":"integer","format":"int64"},"packets":{"type":"array","items":{"$ref":"#/components/schemas/Packet"}}}},"BatchRequest":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"productSkuCode":{"type":"string"},"quantity":{"type":"integer","format":"int32"},"manufacturerDate":{"type":"string","format":"date"},"expirationDate":{"type":"string","format":"date"},"warehouseId":{"type":"integer","format":"int64"},"pageDTO":{"$ref":"#/components/schemas/PageDTO"},"supplierId":{"type":"integer","format":"int64"},"status":{"type":"string"},"inPrice":{"type":"integer"},"outPrice":{"type":"integer"}}},"RfidDTO":{"type":"object","properties":{"rfidCode":{"type":"string"},"batchId":{"type":"integer","format":"int64"},"skuCode":{"type":"string"},"manufacturerDate":{"type":"string","format":"date"},"expirationDate":{"type":"string","format":"date"},"quantity":{"type":"integer","format":"int32"}}},"UserSearchRequest":{"type":"object","properties":{"search":{"type":"string"},"warehouseId":{"type":"integer","format":"int64"},"pageDTO":{"$ref":"#/components/schemas/PageDTO"}}},"ProductLogSearchRequest":{"type":"object","properties":{"action":{"type":"string"},"search":{"type":"string"},"startDateTime":{"type":"string","format":"date-time"},"endDateTime":{"type":"string","format":"date-time"},"status":{"type":"string"},"pageDTO":{"$ref":"#/components/schemas/PageDTO"}}},"ImportRequest":{"type":"object","properties":{"filename":{"type":"string"}}},"ExportLogRequest":{"type":"object","properties":{"skuCode":{"type":"string"},"batchId":{"type":"integer","format":"int64"},"action":{"type":"string"},"startDate":{"type":"string","format":"date-time"},"endDate":{"type":"string","format":"date-time"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}